feat: add SSR to me and home pages
