+++ web-app/parla/app/me/loading.tsx
@@ -1,0 +1,5 @@
+import { MeSkeleton } from "@/components/me/MeSkeleton";
+
+export default function MeLoading() {
+  return <MeSkeleton />;
+}
